refactor(agentic-ai): drop ChatModelProviderRegistry interface
Merge the single-impl interface into one concrete class. The registry
is tightly coupled to ChatModelFactoryImpl and the interface only
existed for mockability in tests, which Mockito handles for concrete
classes.
